게시글 수정 페이지 수정 - 미디어 부분 카테고리에 맞게 수정 - 글목록 링크 제거
글 삭제후 해당 게시판 목록으로 이동
- category를 히든 인풋태그로 넣고, 삭제 전에 컨트롤러에서 redirect 어트리뷰트로 category 보내주기

동영상이나 사진 게시판 별로 필수요소로 만들기
- insertValid에서 카테고리 받아와서 카테고리 별로 유효성 체크 정의

미디어 용량 20mb정도로 제한/ 동영상 용량은 최종편집후 저장되는 용량으로 제한하자
- 서버에서 용량제한은 추천하지 않는 사항이므로 클라이언트에서 20MB로 제한하고 js로 alert표시
- 동영상 용량은 편집전 용량 20MB로 제한됨

동영상 시작 지속을 시작 끝 으로 바꾸고 '끝-시작' 값으로 지속길이 변수를 지정

js에서 onloadedmetadata를 통해 동영상 길이를 구하고 끝에 입력 가능한 최대시간으로 설정

게시글 등록에서 유효성검사시 (시작 > 끝 - 2) 면 false 반환
- 파일이 존재하고 그 타입이 video 일때만 확인하도록

시작 끝 값 손대지 않으면 비디오 편집 기능 생략하도록
- duration 파라미터를 서버로 보내서 끝-시작 == duration 이면 편집안하고 바로 저장하도록

+++ BoardWeb/src/main/java/com/fastwon/video/VideoProcessing.java
@@ -5,7 +5,7 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 
 public class VideoProcessing {
-    public static void trimVideo(String inputPath, String outputPath, int start, int duration) {
+    public static void trimVideo(String inputPath, String outputPath, double start, double duration) {
         try {
             // 현재 디렉토리 경로 얻기
             String currentDirectory = System.getProperty("user.dir");
